+++ app/routes/userRoutes.js
@@ -3,13 +3,11 @@ const UserController = require('../controllers/userController');
 const validateParams = require('../middlewares/validateMiddleware');
 const asyncHandler = require('../utils/asyncHandler');
 
-const { createUserValidatorSchema } = require('../validators/userValidator');
 const { updateUserValidatorSchema } = require('../validators/userValidator');
 
 const router = express.Router();
 
 router.get('/api/v1/users', asyncHandler(UserController.getAllUsers));
-router.post('/api/v1/users', validateParams(createUserValidatorSchema), asyncHandler(UserController.createUser));
 router.get('/api/v1/users/:id', asyncHandler(UserController.getUserById));
 router.patch('/api/v1/users/:id', validateParams(updateUserValidatorSchema), asyncHandler(UserController.updateUser));
 // router.delete('/users/:id', UserController.deleteUser);
