+++ app/services/userService.js
@@ -1,10 +1,7 @@
 const PortalUserModel = require('../models/userModel');
-const sequenceService = require('./sequenceService');
 const CustomError = require('../utils/customError');
 const { STATUS_TYPE } = require('../utils/statusCodes');
-const {hashidsEncode, hashidsDecode} = require('../utils/hashidsHandler');
 const crypto = require('crypto');
-const moment = require('moment');
 
 class UserService {
   async getAllUsers() {
@@ -48,57 +45,6 @@ class UserService {
     }
 
     return queryData;
-  }
-
-  async createUser(name, email, password, refCode='') {
-    const existingUser = await PortalUserModel.findOne({ email }).lean();
-    if (existingUser) {
-      console.log(STATUS_TYPE.PORTAL_EMAIL_ALREADY_REGISTERED)
-      throw new CustomError(STATUS_TYPE.PORTAL_EMAIL_ALREADY_REGISTERED);
-    }
-    const user = new PortalUserModel({ name, email, password});
-
-    // if new user is recommanded by other user, we can get referrer's information
-    if (refCode) {
-      const seqId = await hashidsDecode(refCode);
-      const refUser = await PortalUserModel.findOne({ seq_id: seqId }).lean();
-      if (refUser) {
-        user.inviter = refUser._id;
-      } else {
-        throw new CustomError(STATUS_TYPE.PORTAL_INVALID_INVITATION_CODE);
-      }
-    }
-
-    const pwdObj = await this.generatePassword(user.password);
-    user.salt = pwdObj.salt;
-    user.password = pwdObj.password;
-    user.nick_name = name;
-    user.seq_id = await sequenceService.getNextSequenceValue('portal_user');
-    user.invitation_code = hashidsEncode(user.seq_id);
-
-    // TODO 自动直接激活
-    user.is_activated = true;
-    // 默认新注册用户免费会员的天数
-    user.vip_time_out_at = moment().add(10, 'days').toDate();
-
-    const doc = await user.save();
-
-    if (!doc) {
-      throw new CustomError(STATUS_TYPE.PORTAL_REGISTRATION_FAILED);
-    }
-
-    //const giveToken = await commonConfig.getGiveToken();
-    //const { from, coin, sign } = giveToken;
-    //await assets.sendToken({ from, email: user.email, token: coin, amount: sign, describe: 'signup', invitee: '', invitee_email: '' });
-
-    // TODO 先取消邮件激活功能
-    // const userId = doc._id;
-    // const start = Date.now();
-    // await this.sendActivateEmail({ userId });
-    // ctx.logger.info('新用户注册：\n  用户Id: \t%j\n  用户信息: \t%j\n  请求时长: \t%j ms\n',
-    //   userId, user, Date.now() - start)\
-
-    return doc;
   }
 
   async updateUser(id, userData) {
