+++ app/validators/authValidator.js
@@ -23,4 +23,19 @@ const sendActivateEmailValidatorSchema = {
   }
 };
 
-module.exports = { signinValidatorSchema, signoutValidatorSchema, sendActivateEmailValidatorSchema };
+const createUserValidatorSchema = {
+  name: {
+    notEmpty: { errorMessage: 'Name is required' }
+  },
+  email: {
+    notEmpty: { errorMessage: 'Email is required' },
+    isEmail: { errorMessage: 'Invalid email format' }
+  },
+  password: {
+    notEmpty: { errorMessage: 'Password is required' },
+    isLength: { options: { min: 6 }, errorMessage: 'Password must be at least 6 characters long' }
+  }
+};
+
+module.exports = { signinValidatorSchema, signoutValidatorSchema, sendActivateEmailValidatorSchema, createUserValidatorSchema };
+
